Rename ProfileDropdown folder
